Refactoring package directory
